requirements.txt
PyYAML~=6.0
fastapi~=0.70.0
uvicorn~=0.15.0